Migrate to sq
Migrate to `sq` instead of using the old generate-hu-3. This will
generate an advanced .well-known folder configuration in
`/home/user/output`. From there you're able to move the folder and it's
contents to your final webroot. This is a breaking change from the
original setup which only generated the HU folder and expected the user
to move it from `/root/hu` to it's final reseting place.

Closes #1 as a policy file is generated as part of the `sq wkd` process.

Breaking Changes:
* Output location of hu folder has moved from /root/hu ->
  /home/user/output/.well-known/openpgpkey/<MAIL_DOMAIN>/hu

It is recommended to move the entire `MAIL_DOMAIN` folder contents over
to your webroot if you're using the Direct Method. If you're using
advanced you can move over the entire `.well-known` folder